+++ src/components/example-image/index.js
@@ -8,20 +8,24 @@ import { BemHelper } from '../../helpers/bem-helper'
 // Styling
 const classes = new BemHelper('example-image')
 
-export function ExampleImage(props, context) {
+export function ExampleImage(props) {
 
-  const { data, meta } = props
-  const title = get(meta, 'title', 'Image Block')
+  const { block } = props
+  const title = get(block, 'meta.title', 'Input Block')
+  const value = get(block, 'data.value', '')
 
   return (
     <div {...classes('container')}>
       <h1>{title}</h1>
-      <img alt='Hello World' src={data.value} />
+      <img alt='Hello World' src={value} />
     </div>
   )
 }
 
 ExampleImage.propTypes = {
-  data: PropTypes.objectOf(PropTypes.string).isRequired,
-  meta: PropTypes.objectOf(PropTypes.string).isRequired
+  block: PropTypes.shape({
+    id: PropTypes.number.isRequired,
+    data: PropTypes.objectOf(PropTypes.string).isRequired,
+    meta: PropTypes.objectOf(PropTypes.string).isRequired,
+  }).isRequired,
 }