+++ CHANGES.md
@@ -8,5 +8,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- Basic Editor Components Setup (Editor, EditorBlock and Example Components)
 - Initial Project Setup with Lint, Test and Bundle Setup
+- Basic Editor Components Setup (Editor, EditorBlock and Example Components)